refactor: moves some util files to `utils` folder
